backend/requirements.txt
Django==3.0.6
psycopg2-binary==2.8.5
djangorestframework==3.11.0
markdown==3.2.2
django-filter==2.2.0
googletrans==2.4.0
gunicorn==20.0.4
django-cors-headers==3.3.0
djoser==2.0.3
djangorestframework-simplejwt==4.4.0
